Rework AuthProvider to expose transformInbound and transformOutbound
Overriding wrapNetworkAdapter was brittle & clumsy. With this approach, if an auth provider needs to transform messages on their way in and out, they can override `transform`, which is an object containing `inbound` and `outbound` transformers.

Also, in addition to configuring AuthProvider by extending the base class, you can alternately just insantiate the base class and pass `authenticate`, `transform`, `okToAdvertise`, etc. to the constructor.