Use the same "type" for sync messages in WS client and server
The websocket adapter wraps sync messages in an envelope which looks
like this:

{
  senderId: string,
  targetId: string,
  channelId: string,
  type: string,
  message: UInt8Array,
  broadcast: boolean,
}

The `type` field is common across all message types and is used to
dispatch the message internally. In the `NodeWSServerAdapter` the type
used when _sending_ a sync message was "sync", but the type expected
when receiving a sync message was "message". This is because the
`BrowserWebSocketClientAdapter` was sending sync messages with type
"message" and didn't examine the type field at all when dispatching
received sync messages. This in turn led to some confusion when
implementing interoperable clients.

Modify the `BrowserWebSocketClientAdapter` to send messages with type
"sync" and modify `NodeWSServerAdapter` to accept messages with type of
both "sync" and "message" as sync messages. Thus making the message type
consistent but maintaining backwards compatiblity.