Add an example sync server
We have moved the automerge-repo-sync-server to a separate repository.
This is appropriate as that codebase will now be able to become more
"industrial" with the addition of monitoring and docker containers and
other such cloudy things. However, we would still like to be able to
easily run examples locally without this stuff, and we would also like
to be able to locally experiment with changes to the sync server.
Therefore, add an example sync server in examples/sync-server which can
be used with the other examples but which is explicitly not intended
to be used anywhere else.