StorageSubsystem: don't delete the new chunk when compacting
When compacting a document we first save the new compacted chunk using a
deterministic name for the snapshot chunk and then remove all the chunks
that were used to load the document originally. This logic has a bug
where if the snapshot chunk name has not changed (because the document
has not changed) then we end up removing the snapshot chunk we just
wrote because it was also the name of a chunk we loaded.

Solution: filter the list of chunk names to remove to ensure the newly
written snapshot chunk is not in the list.