Save documents to storage as soon as we create them
Problem: currently it is possible to create a DocHandle, obtain the URl
to that dochandle, and then never save anything to the dochandle (e.g.
by closing the browser window) and therefore never save anything to
storage. This means that you have a URL which points at a document which
can _never_ be resolved. This is not a theoretical problem. In several
example applications we have this pattern of creating a `DocHandle`,
then immediately adding the URL to the window location. Because many of
these applications don't call `DocHandle.change` until the user has
pressed a button (e.g. to incrementa counter) it is possible to refresh
the browser window before any call to `DocHandle.change` and end up in a
situation where the `Repo` is trying to find a document which never
existed.

Solution: Save the document as soon as the handle is created. This might
waste a few bytes of storage but it means we never hand out URLs to
documents which don't exist.