Add a setup phase for network adapters
Problem: `Repo.find` can return a `DocHandle` which is unavailable when the
document is not in storage even if the document _is_ available from
connected peerrs. The issue is that the network subsystem takes some
time to set up connections to it's peers. For example, a
`BrowserWebSocketClientAdapter` starts without a Peer ID from the server.
This means that if the user requests a `DocHandle` whilst we are still
waiting for the server to respond with it's peer ID then the `Repo`
decides that we have no peers and consequently immediately marks the
`DocHandle` as unavailable. This is bad because it's extremely surprising
to the user and requires them to do awkward things like "wait ten
seconds for the websocket to start up".

Solution: Add a setup phase to the `NetworkSubsystem`. Each `NetworkAdapter`
now emits a "ready" event. The `NetworkSubsystem` emits a "ready" event
once all it's adapters have done so. To communicate this to the
`DocHandle` we add a state to the `DocHandle` state machine for "awaiting
network". When the Repo creates a `DocHandle` in `find` it checks if the
`NetworkSubsystem` is ready and if not it puts the handle in the
`AWAITING_NETWORK` state and once the `NetworkSubsystem` tells the
`Repo` it's ready the `Repo` puts the `DocHandle` into the `REQUESTING`
state.