use the `next` API from automerge
The `unstable` API in automerge - which contains a number of desirable
new features such as the Peritext rich text CRDT - has graduated to
`next`. New applications should use the `next` API.

In order to make this easy the latest automerge package publishes a
"subpath export" so that in many cases the only change to make is
changing `import * as A from "@automerge/automerge"` to `import * as A
from @automerge/automerge/next"`. In order to use this though we need to
use the `Node16` `moduleResolution` in our `tsconfig.json`s. This in
turn required a bunch of mechanical updates appending `.js` to local
imports. Therefore, whilst the actual change this commit represents is
small. It unfortunately has to touch a bunch of files.

To summarize:

* Update `@automerge/automerge` to `2.1.0-alpha.12` to get the `/next`
  subpath export
* Update all imports of `@automerge/automerge` to
  `@automerge/automerge/next`
* Update `moduleResolution` in every `tsconfig.json` to `Node16`
* Add `.js` to the end of any local imports