Make repo tests test slow connection setup, again
The Repo tests have a test which checks everything still works in the
event of the other end being slow to respond. In the last commit adding
a setup phase to the network subsystem I broke this logic so it no
longer tested the slow setup case effectively. Reintroduce the original
logic.