fix monorepo dependencies
Fixes https://github.com/automerge/automerge-repo/issues/169

Monorepos are hard. One needs to keep dependency versions and ranges accurate across the repository, and keep checking as the codebase grows. In this PR, I used `@manypkg/cli` (a monorepo "linter") to check and fix versions across the repo, and applied changes as required.

Of note, this fixes the apps under `/examples`.