DocSynchronizer: request unavailable documents from new peers
Problem: There was a missing edge case in the handling of unavailable
DocHandles where the NetworkAdapter was ready but it had not announced
any peers. When a DocHandle was requested which was not available
locally it would be correctly marked as unavailable but when a new peer
arrived on the NetworkAdapter which _did_ have the document the
unavailable document was never requested. This occurred because the
logic in `beginSync` (which is called when a new peer is announced for
every DocHandle for which `sharePolicy` returns true) would immediately
return if the document was unavailabl.e

Solution: modify `beginSync` to check if any of the peers which were
passed to `beginSync` are new and if so send them a sync message even if
the document is locally unavailable.