Add a create helper for node apps
It's frequently useful to be able to quickly jump into a new project
with the websocket network adapter and filesystem storage set up so that
you can start experimenting with repo. This commit adds an NPM
"initializer" which can be used as follows:

   npm create @automerge/node-repo-app

This quickly creates an NPM package with the necessary dependencies and
a simple `index.js`.