Forward sync states across mutliple hops (#233)
Problem: applications often want to know the remote heads of peers who are not directly connected to them. For example, consider this topology

tab -> service worker -> sync server <- service worker <- tab

Here the tab on the left may wish to know whether the tab on the right has all the changes the left has sent, but they are not directly connected so there is no way to ask.

Solution: forward remote heads which we have heard about to interested peers. A peer interested in hearing about heads (let's call them a "listening" peer) can issue a message to the peers they are connected to (let's call them "middle" peers) stating that they wish to be notified of the remote heads state of a set of storage IDs. In our example above the left tab is the listening peer, the service workers and sync servers are middle peers, and the right tab and service worker are using the storage ID the left tab is interested in listening  to.  Whenever a middle peer receives a sync message from one of the storage IDs they send a remote heads update message to

a) Any peers form whom the `sharePolicy`  returns `true`
b) Any listening peers who are connected to the document ID the sync message relates to

The message the middle peers send informs the listening peers of the new remote heads. This message also has a timestamp on it. Middle peers keep track of the last timestamp they received and don't forward remote heads updates which are older than the last timestamp they've seen, which stops messages from looping eternally. 

Finally, middle peers forward remote heads updates using the same logic.

---------

Co-authored-by: Peter van Hardenberg <pvh@pvh.ca>
Co-authored-by: github-actions <github-actions@github.com>
Co-authored-by: Paul Sonnentag <paul.sonnentag@googlemail.com>
Co-authored-by: Alex Currie-Clark <acurrieclark@gmail.com>
Co-authored-by: HerbCaudill <herb@devresults.com>