bugfix: avoid returning stale data in useDocument and useHandle
useDocument and useHandle had a footgun where after the input url changed,
they would continue returning the old doc/handle for the previous input
for 1 or more renders.

This behavior could cause problems in downstream computations or
useEffect hooks which did not account for this brief inconsistency.

This patch fixes the problem by checking whether the input and output
match before returning out of the hook. In useHandle we can just
check the handle URL directly; in useDocument we need to store some
additional state to remember the URL associated with a given doc.