expose load and receive sync message metrics
Problem: when analyzing a sync server under heavy load it is difficult
to identify what is causing the load. There are lots of questions we
might have such as "what documents are in memory?" or "which documents are
taking longest  in `receiveSyncMessage`?". 

Solution: expose a `"doc-metrics"` event on `Repo` which is fired
whenever a documnt is loaded or receives a sync message. This can be
wired up to Prometheus or other monitoring tools to allow operators to
answer these questions.

Add num ops and changes to events