isReadOnly() -- test if a dochandle can be change()d
We currently only support one way of putting a docHandle into readOnly mode, which is to view at particular point in time in the past. This is occasionally requested functionality though, so in preparation for introduction of Keyhive ACLs and other similar work, I've taken the liberty of introducing a new test function here.