Add the ability to export a bundle of documents from a Repo
Problem: there are many usecases where it would be useful to be able to
transfer a set of documents from one repo to another without having a
direct connection between them. Some important usecases are initializing
an application with some shared document history and sync-via-email. At
the moment this is not possible. Even though the underlying automerge
document can be saved to a binary format, automerge-repo provides no
APIs for importing such a binary file into an existing document without
first finding that document somewhere.

The reason we have not added this import feature to date is that
allowing the user to set the document ID for some arbitrary document
is a giant footgun. It allows users to create documents which are
unrelated to each other but have the same ID, which generally results in
extremely surprising merge behavior.

Solution: add a new "bundle" concept. A bundle is a set of document IDs
and their associated automerge documents serialized to a binary
format. A bundle is created by calling `Repo.exportBundle(<list of
DocHandles>)` and then imported into another repo by calling
`Repo.importBundle`. This mechanism means that document IDs are still
created by the repo - which means that there is no risk of a user
accidentally creating two documents with the same ID.

Bundles can be encoded to a binary format using `Bundle.encode` and
decoded using `Bundle.decode` - though this latter function is not
necessary most of the time as you can also pass a `Uint8Array` directly
to `Repo.importBundle`. The binary format is intended to be forwards
compatible so that bundles can be kept around for a long time.