Allow specifying the document ID in Repo.import
Problem: Repo.import allows you to import a raw automerge document into
a repo, but it does not allow you to specify the document ID that the
document hsould be imported as. This makes some usecases such as
asynchronous collaboration over email difficult as Repo provides no way
to obtain a specific document ID except by a direct connection to a peer
who has the document. The reason for this omission is that setting the
document ID directly is a bit of a footgun. It would allow users to
create documents with the same ID but unrelated histories, which can
lead to very unexpected merge behavior (e.g. I create {contacts:
["alice"]}, you create {contacts: ["bob"]} and we both import them into
the same document ID and we end up with either ["alice"] or ["bob"]
where many people would expect ["alice", "bob"]).

Solution: Allow specifying the document ID in Repo.import, but document
it as a bit of a footgun. A later commit introduces an
`automerge-repo-bundles` package which provides a safe way to use this
API.