Add the ability to export a bundle of documents from a Repo
Problem: there are many usecases where it would be useful to be able to
transfer a set of documents from one repo to another without having a
direct connection between them. Some important usecases are initializing
an application with some shared document history and sync-via-email. At
the moment this is not possible. `automerge-repo` allows us to specify
the document ID in Repo.import, but it's easy to misuse this feature and
requires a bunch of boilerplate to ship the document IDs around with the
document data.

Solution: add a new package `automerge-repo-bundles` that provides a
serialization format for a set of document IDs and their associated data
called "bundles" and encapsulates "importing" and "exporting" bundles
from a repo in a safe manner using `exportBundle` and `importBundle`.

Bundles can be encoded to a binary format using `Bundle.encode` and
decoded using `Bundle.decode` - though this latter function is not
necessary most of the time as you can also pass a `Uint8Array` directly
to `importBundle`. The binary format is intended to be forwards
compatible so that bundles can be kept around for a long time.