Avoid re-evaluating useDocHandles hook unnecessarily
We're currently updating the hook's handle map whenever the list of
ids changes, causing it to trigger a re-render. This is intended, but
we are passing the ids Array directly in as a useEffect dependency,
meaning that any usage of the hook with a literal array, like

  useDocHandles([id1, id2, id3])

will create a new array on each render, re-triggering the useEffect
hook, causing a setState and another render, indefinitely.

This code could use some refactoring (it's not obeying React's rules
of hooks), but for now, work around the issue by spreading the id
array into the useEffect dependency array.

Fixes #479.