Fix useDocHandle unnecessary re-renders (#487)
With `suspense: false`, `useDocHandle` would thrash the state by calling `setHandle` even when the `currentHandle` was already available. In some projects, this leads to a dramatic performance degradation as many components re-rendered unnecessarily.

The rules of hooks force us to place hooks before early returns so we just need to short-circuit inside the `useEffect`, like we already do for suspense.