Make state management more explicit
Currently the lifecycle of a document is a little tricky to follow. We
have an xstate state machine in `DocHandle`, but the events which modify
that state machine are emitted by various asynchronous processes which
have a reference to the document scattered throughout the system. This
makes it hard to answer questions like "why is this document
unavailable" because we don't have a single place in the codebase where
we have all the information used to make that decision at the same time.

This commit is an attempt to reorganise state management to make these
kinds of questions easier to answer. Roughly speaking it looks like
this:

* Replace the `CollectionSynchronizer` and `DocSynchronizer` classes
  with a single `DocumentPhasor` class, which manages the lifecycle
  of a single document.
* Move all the state logic out of `DocHandle`. `DocHandle`s are now
  just a place for users to register interest in a document and get
  notified of changes. All actual state changes are routed through
  the `DocumentPhasor`

The "DocumentPhasor" state machine manages a set of `Phase`s, each of
which represents different stages of the document lifecycle ("loading",
"requesting", "ready", "unavailable"). Modifications to the phasor are
achieved by passing events to the `DocumentPhasor.handleEvent` method,
which returns an object describing what has changed.