Add ShareConfig
Problem: the `sharePolicy` field on `RepoConfig` allows configuring
which peers we will announce documents to, but there is no way to stop a
peer who knows a document ID from requesting it from us. We would like
to be able to configure access policies which don't share even with
peers who know the document ID.

Solution: add a new `shareConfig` field which has two fields `announce`
is a function which has the same meaning as the existing `sharePolicy`
field - it determines whether we should announce a document to a peer.
The `access` field is a function which determines whether a peer should
even have access to a document. If `access` returns false we respond to
requests for the docuemnt with a "doc-unavailable" message.