Repo: Pass the initial heads of a document to the id factory
Problem: the purpose of the idFactory callback which we are adding to
`RepoConfig` is to allow keyhive to control the creation of a document
ID, so that the document ID is the same as the keyhive document ID. At
the moment the idFactory is called without any arguments, but the
keyhive document ID requires the initial heads of the document in order
to encode them into the initial delegation of the document creation.

Solution: pass the initial heads of the document to the idFactory. This
is even more of a code smell than the original idFactory, but it is what
we need for keyhive. This is all behind an @hidden flag because we
think we'll probably want to change it anyway.