Add Repo.create2 to allow asynchronous id generation
Problem: in 722a48f9 we added an experimental constructor argument to
Repo which allowed the user to provide a function used to generate new
document IDs. The purpose of this function was so that keyhive can
provide the IDs for documents, which in turn means that we don't have to
carry around a mapping of keyhive IDs to document IDs in keyhive
applications. Unfortunately it turns out that the keyhive ID generation
is asynchronous, which means that we can't call it in Repo.create.

Solution: add a new method Repo.create2 which is asynchronous and which
calls the id generator if present. This method is hidden and
experimental. Long term I think we will need to do a major version
change for Repo and make Repo.create asynchronous, but for now we can
experiment with create2