don't use jsdom in tests
Problem: some tests were failing due to a mysterious issue which
manifested like this in the logs:

    error receiving message {
      err: 'bad theirHave: error loading have at index 0: bad bloom: the value was not a Uint8Array',
      ..
    }

This is an error which is thrown by the wasm-bindgen generated wrapping
code in automerge. In this case the flow is this:

* A sync state is created
* This sync state is passed to a call to `Automerge.receiveSyncMessage`
* Internally, automerge converts the `theirHave` field of the sync state
  to a Vec<u8>. This conversion performs an instanceof check which looks
  roughly like `if (!(theirHave instanceof Uint8Array)) { throw ... }`

The instanceof check was failing. The reason it was failing is that
`jsdom` polyfills the Uint8Array class, so the Uint8Array in the test
environment is not the same as the Uint8Array in the automerge wasm
module.

Solution: don't use jsdom. Instead use happy-dom, which appears to not
polyfill Uint8Array